upgrade JDK to 21
- upgrade JDK from 17 to 21
- fix compile errors with java 25
- fix some compiler warnings

Since JDK 23, the maven compiler plugin needs to have lombok annotations
explicitly enabled.
https://inside.java/2024/06/18/quality-heads-up/

+++ operator/src/main/java/com/datastax/oss/kaap/controllers/BaseResourcesFactory.java
@@ -794,9 +794,9 @@ public abstract class BaseResourcesFactory<T> {
             return false;
         }
 
-        final int replicas = status.getReplicas().intValue();
-        final int ready = status.getReadyReplicas().intValue();
-        final int updated = status.getUpdatedReplicas().intValue();
+        final int replicas = status.getReplicas();
+        final int ready = status.getReadyReplicas();
+        final int updated = status.getUpdatedReplicas();
         return replicas == ready && updated == ready;
     }
